Working prototype
Fixed:
- Missing module in package.json: dxf-parser
- Breaks image loading
- Separate ../reducer/dxf.js file.
- Code formatted to match the existing style, indent=4.
