+++ src/reducers/document.js
@@ -7,6 +7,7 @@ import Snap from 'snapsvg-cjs';
 import { forest, getSubtreeIds, object, reduceParents, reduceSubtree } from '../reducers/object'
 import { addDocument, addDocumentChild } from '../actions/document'
 import { elementToRawPaths, flipY, hasClosedRawPaths } from '../lib/mesh'
+import { processDXF } from '../reducers/dxf'
 
 const initialDocument = {
     id: '',
@@ -161,13 +162,8 @@ function loadImage(state, {file, content}) {
     return state;
 }
 
-// TODO: take out of global scope somehow
-var LayerLookup = new Map();
-
 function loadDxf(state, settings, {file, content}) {
   state = state.slice();
-  LayerLookup = new Map();
-
   let docFile = {
       ...initialDocument,
       id: uuid.v4(),
@@ -178,100 +174,7 @@ function loadDxf(state, settings, {file, content}) {
       selected: false,
   };
   state.push(docFile); // state[0] is the file root structure
-  processDXF(docFile, content);
-
-  function processDXF(docFile, dxfTree) {
-    let i, entity;
-    let newLayer = false;
-    var fileLayers = [];
-
-    let docLayer = {
-        ...initialDocument,
-        type: 'LAYER',
-        isRoot: false,
-        children: [],
-    }
-
-    for(i = 0; i < dxfTree.entities.length; i++) {
-      entity = dxfTree.entities[i];
-      if(entity.type === 'DIMENSION') {
-        console.log('WARNING: No block for DIMENSION entity');
-      } else {
-        // ID layers
-        if (!LayerLookup.get(entity.layer)) {
-          LayerLookup.set(entity.layer, uuid.v4())
-          newLayer = true;
-        }
-        if (newLayer) {
-          docLayer.id = LayerLookup.get(entity.layer);
-          docLayer.name = 'LAYER' + ': ' + entity.layer;
-          state.push(docLayer);
-          docFile.children.push(docLayer.id); // register layer under file
-          drawEntity(entity, dxfTree, docLayer, i);
-          newLayer = false;
-        } else {
-          drawEntity(entity, dxfTree, docLayer, i);
-        }
-        //drawEntity(entity, dxfTree, doc, i);
-      }
-    }
-    //fileLayers = [ ...new Set(fileLayers) ]; // list of unique layer names
-  }
-
-  function drawEntity(entity, dxfTree, docLayer, index) {
-    //console.log('inside drawEntity:  Entity ', entity, '  Index: ', index)
-    if(entity.type === 'CIRCLE' || entity.type === 'ARC') {
-        drawCircle(entity, index);
-    } else if(entity.type === 'LWPOLYLINE' || entity.type === 'LINE' || entity.type === 'POLYLINE') {
-        drawLine(entity, index, docLayer);
-    } else if(entity.type === 'TEXT') {
-        drawText(entity, index);
-    } else if(entity.type === 'SOLID') {
-        drawSolid(entity, index);
-    } else if(entity.type === 'POINT') {
-        drawPoint(entity, index);
-    }
-  }
-
-  function drawLine(entity, index, docLayer) {
-          let docEntity = {
-              ...initialDocument,
-              id: uuid.v4(),
-              type: entity.type,
-              name: entity.layer + ': ' + entity.type + ': ' + entity.handle,
-              isRoot: false,
-              children: [],
-              selected: false,
-          }
-
-          // create geometry
-          let rawPaths = [];
-          for(let j = 0; j < entity.vertices.length; j++) {
-            let p = [];
-            let vertex = {};
-            vertex = entity.vertices[j];
-            p.push(vertex.x, vertex.y);
-            if (p.length)
-              rawPaths = rawPaths.concat(p);
-          }
-          if (entity.shape) {
-            rawPaths = rawPaths.concat(entity.vertices[0].x, entity.vertices[0].y); // To close off the shape
-          }
-
-          if (rawPaths.length) {
-            docEntity.rawPaths = [];
-            docEntity.rawPaths[0] = rawPaths;
-            docEntity.translate = [0, 0, 0];
-            docEntity.scale = [1, 1, 1];
-            docEntity.strokeColor = [0,0,0,0.3];
-            docEntity.fillColor = [0,0,0,0];
-          }
-
-          state.push(docEntity);
-          docLayer.children.push(docEntity.id); // register feature under layer
-
-  }
-
+  processDXF(state, docFile, content);
   return state;
 }
